Refactor and add functions to retrieve topics, comment some future features not yet implemented in core API
