--- src/default-list-params.ts
@@ -1,11 +1,0 @@
-export default {
-  dateFrom: 'now-7d',
-  dateTo: 'now',
-  langs: [],
-  products: [],
-  query: '',
-  minDocCount: 1,
-  urgencies: [],
-  sources: [],
-  topics: []
-}
